+++ examples/crd_api.rs
@@ -11,6 +11,7 @@ use k8s_openapi::apiextensions_apiserver::pkg::apis::apiextensions::v1beta1 as a
 
 use kube::{
     api::{Api, DeleteParams, ListParams, Patch, PatchParams, PostParams, ResourceExt},
+    core::crd::v1beta1::CustomResourceExt,
     Client, CustomResource,
 };