+++ kube-core/src/params.rs
@@ -287,15 +287,15 @@ impl PatchParams {
         Ok(())
     }
 
-    pub(crate) fn populate_qp(&self, qp: &mut url::form_urlencoded::Serializer<String>) {
+    pub(crate) fn populate_qp(&self, qp: &mut form_urlencoded::Serializer<String>) {
         if self.dry_run {
             qp.append_pair("dryRun", "All");
         }
         if self.force {
             qp.append_pair("force", "true");
         }
-        if let Some(ref field_manager) = self.field_manager {
-            qp.append_pair("fieldManager", &field_manager);
+        if let Some(ref fm) = self.field_manager {
+            qp.append_pair("fieldManager", fm);
         }
     }